refactor(tests): split EarmarkBudgetTests (458→199) (#272)
Move the `BudgetLineItem Merge` suite (pure model logic, no store fixture)
into a sibling BudgetLineItemMergeTests.swift file. `EarmarkStore Budget`
(store-level integration tests against TestBackend) stays in the original.

The two suites were already independent — they share no helpers and each
tests a different layer — so the seam is clean.

Drops the file_length baseline entry entirely; no other baselined rules
were keyed to this file. No access widenings (makeStore stays private).
